requirements.txt
lz4
sanic
requests
websocket-client
